requirements.txt
appium-python-client>=3.0
selenium>=4.0
requests>=2.31.0
